move replacevalue to utils
