+++ lib/command/run-multiple.js
@@ -8,6 +8,7 @@ const runHook = require('../hooks');
 const event = require('../event');
 const collection = require('./run-multiple/collection');
 const clearString = require('../utils').clearString;
+const replaceValue = require('../utils').replaceValue;
 
 const runner = path.join(__dirname, '/../../bin/codecept');
 let config;
@@ -184,25 +185,4 @@ function executeRun(runName, runConfig) {
     .on('error', (err) => {
       return onProcessEnd(1);
     });
-}
-
-
-/**
- * search key in object recursive and replace value in it
- */
-function replaceValue(obj, key, value) {
-  if (!obj) return;
-  if (obj instanceof Array) {
-    for (const i in obj) {
-      replaceValue(obj[i], key, value);
-    }
-  }
-  if (obj[key]) obj[key] = value;
-  if (typeof obj === 'object' && obj !== null) {
-    const children = Object.keys(obj);
-    for (let childIndex = 0; childIndex < children.length; childIndex++) {
-      replaceValue(obj[children[childIndex]], key, value);
-    }
-  }
-  return obj;
 }
