+++ lib/utils.js
@@ -305,3 +305,23 @@ module.exports.beautify = function (code) {
   const format = require('js-beautify').js;
   return format(code, { indent_size: 2, space_in_empty_paren: true });
 };
+
+/**
+ * search key in object recursive and replace value in it
+ */
+module.exports.replaceValue = function replaceValue(obj, key, value) {
+  if (!obj) return;
+  if (obj instanceof Array) {
+    for (const i in obj) {
+      replaceValue(obj[i], key, value);
+    }
+  }
+  if (obj[key]) obj[key] = value;
+  if (typeof obj === 'object' && obj !== null) {
+    const children = Object.keys(obj);
+    for (let childIndex = 0; childIndex < children.length; childIndex++) {
+      replaceValue(obj[children[childIndex]], key, value);
+    }
+  }
+  return obj;
+}
